requirements.txt
# Web scraping
playwright>=1.41.0
html5lib>=1.1

# Search engine
duckduckgo-search>=7.2.1

# LLM integration
openai>=1.59.8 # o1 support
anthropic>=0.42.0
python-dotenv>=1.0.0

# Testing
unittest2>=1.1.0
pytest>=8.0.0
pytest-asyncio>=0.23.5

# Google Generative AI
google-generativeai

# gRPC, for Google Generative AI preventing WARNING: All log messages before absl::InitializeLog() is called are written to STDERR
grpcio==1.70.0

# EU AI Act Compliance Tool dependencies
langchain>=0.1.0
langchain_openai>=0.1.0
pymongo>=4.6.1
pinecone>=3.0.0
pypdf>=4.0.1
unstructured>=0.11.0
tiktoken>=0.5.0
python-docx>=1.0.0
matplotlib>=3.8.0
plotly>=5.18.0
pandas>=2.1.0
